# specs/001-backend-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Task Management API
  description: REST API for managing user tasks with JWT authentication
  version: 1.0.0
  contact:
    name: Todo Full-Stack Project

servers:
  - url: http://localhost:8000
    description: Local development server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token issued by Better Auth (frontend)

  schemas:
    TaskStatus:
      type: string
      enum: [pending, in_progress, completed]
      default: pending

    Priority:
      type: string
      enum: [low, medium, high]
      default: medium

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 100
          minLength: 1
          description: Task title (required)
        description:
          type: string
          maxLength: 500
          nullable: true
          description: Task description (optional)
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/Priority'

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
          minLength: 1
        description:
          type: string
          maxLength: 500
          nullable: true
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        is_completed:
          type: boolean

    TaskResponse:
      type: object
      required:
        - id
        - user_id
        - title
        - status
        - priority
        - is_completed
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique task identifier
        user_id:
          type: string
          description: Owner's user ID
        title:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
          nullable: true
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        is_completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    TaskList:
      type: array
      items:
        $ref: '#/components/schemas/TaskResponse'

    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message

    ValidationError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string

  responses:
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing_token:
              value:
                detail: "Missing authorization header"
            invalid_token:
              value:
                detail: "Invalid token"
            expired_token:
              value:
                detail: "Token has expired"
            user_mismatch:
              value:
                detail: "User ID mismatch"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Task not found"

    ValidationFailed:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'

security:
  - BearerAuth: []

paths:
  /api/{user_id}/tasks:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
        description: User ID (must match JWT sub claim)

    get:
      summary: List all tasks for user
      operationId: getTasks
      tags:
        - Tasks
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create a new task
      operationId: createTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'

  /api/{user_id}/tasks/{task_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
        description: User ID (must match JWT sub claim)
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Task UUID

    get:
      summary: Get a single task
      operationId: getTask
      tags:
        - Tasks
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update a task
      operationId: updateTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'

    delete:
      summary: Delete a task
      operationId: deleteTask
      tags:
        - Tasks
      responses:
        '204':
          description: Task deleted (no content)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/{user_id}/tasks/{task_id}/complete:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
        description: User ID (must match JWT sub claim)
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Task UUID

    patch:
      summary: Toggle task completion status
      operationId: toggleTaskComplete
      tags:
        - Tasks
      responses:
        '200':
          description: Task completion toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

tags:
  - name: Tasks
    description: Task CRUD operations